Rust scoring code uses external libc
In order to get this ready for a 1.0 release, I updated it to use the
libc from crates.io as that is not listed as unstable. This means I had
to start using Cargo.

I left the benchmarking code in, but commented it out. This is not
stable and will not work on the upcoming 1.0 release. I will be
benchmarking on nightly releases.